test: split Project Intake workflow coverage
